% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cebq_score.R
\name{cebq_score}
\alias{cebq_score}
\title{cebq_score: Calculates the Children's Eating Behavior Questionnaire sub-scales for PROPERLY formatted data}
\usage{
cebq_score(cebq_dat, parID_var, reverse_scored)
}
\arguments{
\item{cebq_dat}{A dataset with the 35 Children's Eating Behavior Questionnaire questions and raw responses with each row representing a participant}

\item{parID_var}{variable name for column containg participant IDs, if cebq_dat contains IDS}

\item{reverse_scored}{(logic) TRUE/FLASE - raw responses have already been reversed scored (e.g., due to how the levels were coded in RedCap)}
}
\value{
A numeric value for each of the eight subscales for each row in the dataset
}
\description{
This function calculates the the eight Children's Eating Behavior Questionnaire subscales:
}
\examples{

\dontrun{
}

}
\seealso{
To get best fit parameters use \code{\link{LODE_Fit}}.
To get fit your intake data using the  the Quadratic model
(Kissileff, 1982; Kissileff & Guss, 2001), see \code{\link{Quad_Fit}}
and \code{\link{Quad_n2ll}}.
}
