% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbs_visit_databases.R
\name{fbs_visit_databases}
\alias{fbs_visit_databases}
\title{fbs_visit_databases: Generate databases for the Food and Brain Study based on study visit}
\usage{
fbs_visit_databases(
  visit1 = FALSE,
  visit2 = FALSE,
  visit3 = FALSE,
  visit4 = FALSE,
  visit5 = FALSE,
  visit6 = FALSE,
  visit7 = FALSE,
  write_dat = TRUE,
  write_path,
  data_path,
  model_DD = FALSE,
  child_file_pattern,
  parent_file_pattern,
  visit_file_pattern
)
}
\arguments{
\item{visit1}{(optional) logical indicator if want visit 1 data in database. Default = FALSE.}

\item{visit2}{(optional) logical indicator if want visit 2 data in database. Default = FALSE.}

\item{visit3}{(optional) logical indicator if want visit 3 data in database. Default = FALSE.}

\item{visit4}{(optional) logical indicator if want visit 4 data in database. Default = FALSE.}

\item{visit5}{(optional) logical indicator if want visit 5 data in database. Default = FALSE.}

\item{visit6}{(optional) logical indicator if want visit 6 data in database. Default = FALSE.}

\item{visit7}{(optional) logical indicator if want visit 7 data in database. Default = FALSE.}

\item{write_dat}{indicate whether to write databases. Default is TRUE.}

\item{write_path}{(optional) a string with the path indicating where to save the generated databases if write_dat is TRUE (default option). If no path is given, databases will be written to working directory.}

\item{data_path}{(optional) the full path to the directory where the raw Qualtrics database are saved, EXCLUDING the .sav file name (e.g., '.../b-childfoodlab_Shared/Active_Studies/RO1_Brain_Mechanisms_IRB_5357/Participant_Data/untouchedRaw/Qualtrics_Raw/'). If NOT entered, will assume the database is saved in the working directory.}

\item{model_DD}{If visit3 is set to TRUE, indicate if delay discounting data should be modeled. This will take an addition 3-5 minutes of processing time. Default = FALSE.}

\item{child_file_pattern}{(optional) This is only needed if file naming deviates from the standard naming: e.g., Child_V#_*sav'. Only enter the string indicating a respondent (e.g., 'Child').}

\item{parent_file_pattern}{(optional) This is only needed if file naming deviates from the standard naming: e.g., Parent_V#_*sav'. Only enter the string indicating a respondent (e.g., 'Child').}

\item{visit_file_pattern}{(optional) This is only needed if file naming deviates from the standard naming for visits: e.g., V# . Only enter the string indicating a visit (e.g., 'V' for 'V1' or 'Visit_' for 'Visit_1').}
}
\value{
A list containing all databases that were generated
}
\description{
This function will generate up-to-date databases for the Food and Brain Study based on visit only. If multiple visits are entered, the data will be merged into a signle database
}
\details{
To process the raw data, the raw databases from Qualtrics MUST follow the naming convention: Child_V1_YYYY-MM-DD.sav, Child_V1_Home_YYY-MM-DD.sav, Child_V1_Lab_YYY-MM-DD.sav, and Parent_V1_YYY-MM-DD.sav. The databases must all be in the SAME directory to be processed if the data_path is not entered and the directory organization does not follow the structure laid out in the DataManual.
}
\examples{
#if in same working directory as data, want delay discounting data modeled:
fbs_data_proc <- fbs_visit_databases(visit3 = TRUE, model_DD = TRUE)

#if only want the full visit 1 database returned to the terminal:
fbs_data_proc <- fbs_visit_databases(visit1 = TRUE, write_dat = FALSE)

#if want the visit 1 and visit 6 data written to the working directory:
fbs_data_proc <- fbs_visit_databases(visit1 = TRUE, visit6 = TRUE)

\dontrun{

}

}
\seealso{
Raw data from Qualtrics is processed using the following scripts: \code{\link{util_fbs_merge_v1dat}}, \code{\link{util_fbs_merge_v2dat}}, \code{\link{util_fbs_merge_v3dat}}, \code{\link{util_fbs_merge_v4dat}}, \code{\link{util_fbs_merge_v5dat}}, \code{\link{util_fbs_merge_v6dat}}, \code{\link{util_fbs_merge_v7dat}}
}
