% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_child_v1dat_home.R
\name{qualtrics_child_v1dat_home}
\alias{qualtrics_child_v1dat_home}
\title{qualtrics_child_v1dat_home: Process raw qualtrics visit 1 home data for the child}
\usage{
qualtrics_child_v1dat_home(date_str, data_path)
}
\arguments{
\item{date_str}{the date used in the name of the .sav file (e.g., for file 'Child_V1_Home_2021-10-11.sav',
the string '2021-10-11' would be entered)}

\item{data_path}{(optional) the full path to the V1 Child Qualtircs database EXCLUDING /Final_CovidAtHome/ and the .sav file name (e.g., '.../b-childfoodlab_Shared/Active_Studies/RO1_Brain_Mechanisms_IRB_5357/Participant_Data/untouchedRaw/Qualtrics_Raw/').}
}
\value{
A list containing: 1) data: data.frame with raw, cleaned data from child visit 1 Qualtrics
collected in the child's home and 2) dict: all variable descriptions
}
\description{
This function loads the .sav raw data file for the child visit 1 that was
collected via Qualtrics at the child's home when the procedure was split due to covid
and cleans the data. Cleaning the data involves:
1) extracting all variable descriptions,
2) selecting relevant data columns,
3) removing all practice events (e.g., 999)
4) re-ordering and re-name data columns
5) reformatting dates to be appropriate and computer readable: YYYY-MM-DD
6) re-calculate manual variables
7) re-ordering factor levels to start with value 0
8) random fixes to factor level names and variable descriptions
}
\examples{
#if in same working directory as data:
ch_v1_dat <- qualtrics_child_v1dat_home('2021-10-11')

\dontrun{
#date must be a string. The following will not run:
ch_v1_dat <- qualtrics_child_v1dat_home(2021-10-11)

#date must match the file name - for file named 'Child_V1_Home_2021_09_16', the
following will not run:
ch_v1_dat <- qualtrics_child_v1dat_home('2021_10_11')
}


}
