% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_child_v4dat.R
\name{qualtrics_child_v4dat}
\alias{qualtrics_child_v4dat}
\title{qualtrics_child_v4dat: Process raw qualtrics visit 4 data for the child}
\usage{
qualtrics_child_v4dat(date_str, data_path)
}
\arguments{
\item{date_str}{the date used in the name of the .sav file (e.g., for file 'Child_V4_2021-10-11.sav',
the string '2021-10-11' would be entered)}

\item{data_path}{(optional) the full path to the V4 Child Qualtircs database EXCLUDING the .sav file name (e.g., '.../b-childfoodlab_Shared/Active_Studies/RO1_Brain_Mechanisms_IRB_5357/Participant_Data/untouchedRaw/Qualtrics_Raw/').}
}
\value{
A list containing: 1) data: data.frame with raw, cleaned data from child visit 4 Qualtrics
and 2) dict: all variable descriptions
}
\description{
This function loads the .sav raw data file for the child visit 4 data that was
collected via Qualtrics and cleans the data. Cleaning the data involves:
1) extracting all variable descriptions,
2) selecting relevant data columns,
3) removing all practice events (e.g., 999)
4) re-ordering and re-name data columns
5) reformatting dates to be appropriate and computer readable: YYYY-MM-DD
6) re-calculate manual variables
7) fix 99's / create prefer not to answer database
8) random fixes to factor level names and variable descriptions
}
\examples{
#if in same working directory as data:
ch_v4_dat <- qualtrics_child_v4dat('2021-10-11')

\dontrun{
#date must be a string. The following will not run:
ch_v4_dat <- qualtrics_child_v4dat(2021-10-11)

#date must match the file name - for file named 'Child_V4_2021_09_16', the
following will not run:
ch_v4_dat <- qualtrics_child_v4dat('2021_10_11')
}


}
