% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_parent_v6dat.R
\name{qualtrics_parent_v6dat}
\alias{qualtrics_parent_v6dat}
\title{qualtrics_parent_v6dat: Process raw Qualtrics visit 6 data for the parent}
\usage{
qualtrics_parent_v6dat(date_str, data_path)
}
\arguments{
\item{date_str}{the date used in the name of the .sav file (e.g., for file 'Parent_V6_2021-10-11.sav',
the string '2021-10-11' would be entered)}

\item{data_path}{(optional) the full path to the visit 3 Parent Qualtrics database EXCLUDING the .sav file name (e.g., '.../b-childfoodlab_Shared/Active_Studies/RO1_Brain_Mechanisms_IRB_5357/Participant_Data/untouchedRaw/Qualtrics_Raw/').}
}
\value{
A list containing: 1) data: data.frame with raw, cleaned data from parent visit 6 Qualtrics;
 2) dict: all variable descriptions; 3) pna_data: data.frame marking participants who 'prefered not to answer' (pna) specific questions; and 4) pna_dict: all variable descriptions for pna_data
}
\description{
This function loads the .sav raw data file for the parent visit 6 data that was
collected via Qualtrics and cleans the data. Cleaning the data involves:
1) extracting all variable descriptions,
2) selecting relevant data columns,
3) removing all practice events (e.g., 999)
4) re-ordering and re-name data columns
5) general fixes to variable labels (remove ' - 1')
6) fix variables with 99 issue for 'prefer not to answer'
7) reformatting dates to be appropriate and computer readable: YYYY-MM-DD
}
\examples{
#if in same working directory as data:
p_v6_dat <- qualtrics_parent_v6dat('2021-10-11')

\dontrun{
#date must be a string. The following will not run:
p_v6_dat <- qualtrics_parent_v6dat(2021-10-11)

#date must match the file name - for file named 'Parent_V6_2021_09_16', the
following will not run:
p_v6_dat <- qualtrics_parent_v6dat('2021_10_11')
}


}
