% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_brief2_lookup.R
\name{ref_brief2_lookup}
\alias{ref_brief2_lookup}
\title{ref_brief2_lookup: Lookup function the Behavioral Rating Inventory of Executive Function-2}
\usage{
ref_brief2_lookup(sex, age, value, item)
}
\arguments{
\item{sex}{numeric value for sex: 1 for male and 0 for female. equired for the following for subscale, index, and GEC values. Not needed for quality control scales (i.e., negativity, inconsistency, and frequency)}

\item{age}{numeric value for age. Required for the following for subscale, index, and GEC values. Not needed for quality control scales (i.e., negativity, inconsistency, and frequency)}

\item{value}{raw value/score to look up}

\item{item}{string matching the item/subscale prefix in lookup tables (e.g., to look up t-score for the BRIEF-2 inhibition subscale, enter 'inhibit')}
}
\value{
A list with values from the lookup table
}
\description{
This function pulls t-scores and percentiles based on age and sex for the the Behavioral Rating Inventory of Executive Function-2
}
\examples{

# look up t-score and percentile for the inhibit subscale of BRIEF-2 for a 7-year-old male with a raw score of 14
brief_score_data <- ref_brief2_lookup(value = 14, age = 7, sex = 0, item = 'inhibit')

\dontrun{
}

}
\seealso{
This can be called from general script: \code{\link{ref_lookup_fn}}
}
