% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_lookup_fn.R
\name{ref_lookup_fn}
\alias{ref_lookup_fn}
\title{ref_lookup_fn: Reference documentation for lookup functions called in other scripts}
\usage{
ref_lookup_fn(ref_table, value, age, sex, item)
}
\arguments{
\item{ref_table}{string matching the prefix of the desired reference table (e.g., 'breif2' will find the brief2_scoretables.rda reference data)}

\item{value}{raw value/score to look up}

\item{age}{numeric value for age. Required for the following ref_tables: 'brief2'}

\item{sex}{numeric value for sex: 1 for male and 0 for female. Required for the following ref_tables: 'brief2'}

\item{item}{string matching the item/subscale prefix in lookup tables (e.g., to look up t-score for the BRIEF-2 inhibition subscale, enter 'inhibit')}
}
\value{
A list with values from the lookup table
}
\description{
This function includes sub-functions that perform table lookup tasks. The included functions include: ref_brief2_lookup
}
\examples{

# look up t-score and percentile for the inhibit subscale of BRIEF-2 for a 7-year-old male with a raw score of 14
brief_score_data <- ref_lookup_fn(brief2, value = 14, age = 7, sex = 0, item = 'inhibit')

\dontrun{
}

}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{qualtrics_parent_v4dat}}
}
