% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_audit.R
\name{score_audit}
\alias{score_audit}
\title{score_audit: Score data from the Alcohol Use Disorders Identification Test}
\usage{
score_audit(audit_data, parID)
}
\arguments{
\item{audit_data}{a data.frame all items for the Alcohol Use Disorders Identification Test following the naming conventions described above}

\item{parID}{(optional) name of participant ID column in intake_data. If included the output dataset will be matched by parID, if not included the output dataset will be in the order of intake_data but will have no participant identifier.}
}
\value{
A dataset with a score for the Alcohol Use Disorders Identification Test
}
\description{
This function scores the Alcohol Use Disorders Identification Test
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'audit#' where # is the question number (1-10)
3) All questions must have the numeric value for the choice:
3a) question 1: 0 - Never, 1 - Monthly or Less, 2 - 2-4 times a month, 3 - 2-3 times a week, 4 - 4 or more times a week
3b) question 2: 0 - 1 or 2, 1 - 3 or 4, 2 - 5 or 6, 3 - 7 to 9, 4 - 10
3c) question 3-8: 0 - Never, 1 - Less than Monthly, 2 - Monthly, 3 - Weekly, 4 - Daily or Almost Daily
3d) questions 9-10: 0 - No, 2 - Yes, but not in the last year, 4 - Yes, during the last year

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the audit with IDs
audit_score_data <- score_audit(audit_data, parID = 'ID')

\dontrun{
}

}
\references{
Saunders JB, Aasland OG, Babor TF, De La Fuente JR, Grant M. Development of the Alcohol Use Disorders Identification Test (AUDIT): WHO Collaborative Project on Early Detection of Persons with Harmful Alcohol Consumption-II. Addiction. 1993;88(6):791-804. doi:10.1111/j.1360-0443.1993.tb02093.x (\href{https://pubmed.ncbi.nlm.nih.gov/8329970/}{PubMed})
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_parent_v5dat}}
}
