% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_bisbas.R
\name{score_bisbas}
\alias{score_bisbas}
\title{score_bisbas: Scored data from the Behavioral Inhibition System/Behavioral Activation System}
\usage{
score_bisbas(bisbas_data, parID)
}
\arguments{
\item{bisbas_data}{a data.frame all items for the Behavioral Inhibition Scale/Behavioral Activation Scale following the naming conventions described above}
}
\value{
A dataset with subscale scores for the Behavioral Inhibition Scale/Behavioral Activation Scale
}
\description{
This function scores the Behavioral Inhibition System (BIS)/Behavioral Activation System (BAS) and provides subscale scores for the following behaviors: BIS, BAS Fun Seeking, BAS Drive, and BAS Reward Responsiveness. Note, this script is used to score the 24-item version, which contains 4 filler questions that are not used for scoring. While the scored questions match the 20-item version exactly, the question numbers differ so this script cannot be sued to score the 20-item version at this time. If this functionality is desired, contact the kellertools package developers.
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'bisbas#' where # is the question number (1-24).
3) Questions must have the numeric value for the choices: 1 - Very True for Me, 2 - Somewhat True for Me, 3 - Somewhat False for Me, 4 - Very False for Me

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the bisbas with IDs
bisbas_score_data <- score_bisbas(bisbas_data, parID = 'ID')


\dontrun{
}

}
\references{
Primary References for the child version of the Behavioral Inhibition Scale/Behavioral Activation Scale and Scoring:
Muris P, Meesters C, de Kanter E, Timmerman PE. Behavioural inhibition and behavioural activation system scales for children: relationships with Eysenck’s personality traits and psychopathological symptoms. Personality and Individual Differences. 2005;38(4):831-841. doi:10.1016/j.paid.2004.06.007

Carver CS, White TL. Behavioral inhibition, behavioral activation, and affective responses to impending reward and punishment: The BIS/BAS Scales. Journal of Personality and Social Psychology. 1994;67(2):319-333. doi:http://dx.doi.org/10.1037/0022-3514.67.2.319.
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_parent_v3dat}}
}
