% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cbis.R
\name{score_cbis}
\alias{score_cbis}
\title{score_cbis: Score data from the Children's Body Image Scale}
\usage{
score_cbis(cbis_data, parID)
}
\arguments{
\item{cbis_data}{a data.frame all items for the Children's Body Image Scale following the naming conventions described above}

\item{parID}{(optional) name of participant ID column in intake_data. If included the output dataset will be matched by parID, if not included the output dataset will be in the order of intake_data but will have no participant identifier.}
}
\value{
A dataset with a score for the Children's Body Image Scale
}
\description{
This function scores the Children's Body Image Scale
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'cbis_perc_x' and 'cbis_ideal_x' where x is male or female
3) All questions must have the numeric value for the choice: 1- Picture A, 2 - Picture B, 3 - Picture C, 4 - Picture D, 5 - Picture E, 6 - Picture F, 7- Picture G

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the cbis with IDs
cbis_score_data <- score_cbis(cbis_data, parID = 'ID')

\dontrun{
}

}
\references{
Truby H, Paxton SJ. Development of the Children’s Body Image Scale. British Journal of Clinical Psychology. 2002;41(2):185-203. doi:10.1348/014466502163967 (\href{https://pubmed.ncbi.nlm.nih.gov/12034005/}{PubMed})
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_child_v4dat}} and \code{\link{util_fbs_child_v4dat_home}}
}
