% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cchip.R
\name{score_cchip}
\alias{score_cchip}
\title{score_cchip: Scored data from the Community Childhood Hunger Identification Project}
\usage{
score_cchip(cchip_data, parID)
}
\arguments{
\item{cchip_data}{a data.frame all items for the Community Childhood Hunger Identification Project following the naming conventions described above}
}
\value{
A dataset with scores for the Community Childhood Hunger Identification Project
}
\description{
This function scores the Community Childhood Hunger Identification Project and provides an overall CCHIP Score and Food Insecurity Status
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'cchip#' where # is question number. For the Food and Brain Study, the 8 primary questions needed are: cchip1, cchip5, cchip9, cchip13, cchip17, cchip21, cchip25, and cchip29. Can be adapted to processes other data/studies in the future.
3) The primary indicatory questions must have the following numeric values: Yes - 1, No - 0

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the cchip with IDs
cchip_score_data <- score_cchip(cchip_data, parID = 'ID')


\dontrun{
}

}
\references{
Wehler CA, Scott RI, Anderson JJ. The community childhood hunger identification project: A model of domestic hunger—Demonstration project in Seattle, Washington. Journal of Nutrition Education. 1992;24(1):29S-35S. doi:10.1016/S0022-3182(12)80135-X
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{qualtrics_parent_v4dat}}
}
