% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cfq.R
\name{score_cfq}
\alias{score_cfq}
\title{score_cfq: Scored data from the Child Feeding Questionnaire}
\usage{
score_cfq(cfq_data, restriction_split = FALSE, study = "fbs", parID)
}
\arguments{
\item{cfq_data}{a data.frame all items for the Child Feeng Questionnaire following the naming conventions described above}

\item{restriction_split}{a boolean incating if the Restriction subscale should be split to remove food as reward items. Default = FALSE. The standard Restriction subscale will always be available. If restriction_split = TRUE, then two adtion scales will be computed: 1) cfq_rest_noreward: questions 17-20, 23-24 and 1) cfq_foodreward: questions 21-22}

\item{study}{a string incating which study collected the data. Currently, only option and default is 'fbs'. This parameter is included so this script can be adapted for future stues that have wider age rang and may collect question 13.}
}
\value{
A dataset with subscale scores for the Child Feeding Questionnaire
}
\description{
This function scores the Child Feeding Questionnaire and provides subscale scores for the following behaviors: Perceived Responsibility, Perceived Child Weight, Perceived Parent Weight, Child Weight Concerns, Restriction, Pressure to Eat, and Monitoring
}
\details{
To use this function, the data must be prepared accordinng to the following criteria:
1) The data must include all invidividual questionnaire items
2) The  columns/variables must match the following naming convention: 'cfq#' where # is the question number (1-31). For the Food Behavior Study, question was 13 skipped for (subscale - Perceived Child Weight) due to age range.
3) All questions must have the numeric value for the choice:
3a) Perceived Weight: 1 - Markedly Underweight, 2 - Underweight, 3 - Average, 4 - Overweight, 5 - Markedly Overweight
3b) Child Weight Concern: 1 - Unconcerned, 2 - Slightly Unconcerned, 3 - Neutral, 4 - Slightly Concerned, 5 - Very Concerned
3c) Restriction and Pressure to Eat: 1 - Disagree, 2 - Slightly Disagree, 3 - Neutral, 4 - Slightly Agree, 5 - Agree
3d) Monitoring: 1 - Never, 2 - Rarely, 3 - Sometimes, 4 - Mostly, 5 - Always

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the cfq with IDs
cfq_score_data <- score_cfq(cfq_data, parID = 'ID')

# scoring for the cfq with extra Restriction subscales
cfq_score_data <- score_cfq(cfq_data, restriction_split = TRUE, parID = 'ID')

\dontrun{
}

}
\references{
Birch, L. L., Fisher, J. O., Grimm-Thomas, K., Markey, C. N., Sawyer, R., & Johnson, S. L. (2001). Confirmatory factor analysis of the Child Feeng Questionnaire: A measure of parental attitudes, beliefs and practices about child feeng and obesity proneness. Appetite, 36(3), 201–210. https://doi.org/10.1006/appe.2001.0398 (\href{https://pubmed.ncbi.nlm.nih.gov/11358344/}{PubMed})
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_parent_v2dat}}
}
