% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ctc.R
\name{score_ctc}
\alias{score_ctc}
\title{score_ctc: Scored data from the Communities that Care}
\usage{
score_ctc(ctc_data, study, parID)
}
\arguments{
\item{ctc_data}{a data.frame all items for the Child Behavior Questionnaire following the naming conventions described above}

\item{study}{a string indicating which study collected the data. Currently, only option and default is 'fbs'. This parameter is included because the Food and Brain study used a highly adapted verison of the Communities that Care Questionnaire with only 16 questions}

\item{parID}{(optional) name of participant ID column in intake_data. If included the output dataset will be matched by parID, if not included the output dataset will be in the order of intake_data but will have no participant identifier.}
}
\value{
A dataset with subscale scores for the Child Behavior Questionnaire
}
\description{
This function scores the Child Behavior Questionnaire and provides subscale scores for the following behaviors: Child Approval of Drug/Alcohol Use, Child View of Drug/Alcohol Use as Harmful, Friend Approval of Drug/Acohol Use, and Parent Disapproval of Drug/Alcohol Use.
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'ctc#' where # is the question number (1-16 for the Food and Brain Study)
3) All questions must have the numeric value for the choice: 1 - Not at all, 2 - A little, 3 - Not sure/in the middle, 4 - Somewhat, 5 - A lot, -99 - Skip

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the ctc with IDs
ctc_score_data <- score_ctc(ctc_data, parID = 'ID')

\dontrun{
}

}
\references{
NEED
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_child_v5dat}}
}
