% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cwc.R
\name{score_cwc}
\alias{score_cwc}
\title{score_cwc: Score data from the Child Weight Concerns}
\usage{
score_cwc(cwc_data, parID)
}
\arguments{
\item{cwc_data}{a data.frame all items for the Child Weight Concerns following the naming conventions described above}
}
\value{
A dataset with a score for the Child Weight Concerns
}
\description{
This function scores the Child Weight Concerns
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'cwc#' where # is the question number (1-5)
3) All questions must have the numeric value for the choice starting with the value 1. Each item is scored differently so responses are scaled to be 0-100 for each item.

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the cwc with IDs
cwc_score_data <- score_cwc(cwc_data, parID = 'ID')

\dontrun{
}

}
\references{
Killen JD, Taylor CB, Hayward C, et al. Pursuit of thinness and onset of eating disorder symptoms in a community sample of adolescent girls: A three-year prospective analysis. Int J Eat Disord. 1994;16(3):227-238. doi:10.1002/1098-108X(199411)16:3<227::AID-EAT2260160303>3.0.CO;2-L (\href{https://pubmed.ncbi.nlm.nih.gov/7833956/}{PubMed})

Taylor CB, Sharpe T, Shisslak C, et al. Factors associated with weight concerns in adolescent girls. Int J Eat Disord. 1998;24(1):31-42. doi:10.1002/(SICI)1098-108X(199807)24:1<31::AID-EAT3>3.0.CO;2-1 (\href{https://pubmed.ncbi.nlm.nih.gov/9589309/}{PubMed})
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{qualtrics_child_v4dat}} and \code{\link{qualtrics_child_v4dat_home}}
}
