% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ffbs.R
\name{score_ffbs}
\alias{score_ffbs}
\title{score_ffbs: Scored data from the Family Food Behvior Survey}
\usage{
score_ffbs(ffbs_data, parID)
}
\arguments{
\item{ffbs_data}{a data.frame all items for the Family Food Behvior Survey following the naming conventions described above}

\item{parID}{(optional) name of participant ID column in intake_data. If included the output dataset will be matched by parID, if not included the output dataset will be in the order of intake_data but will have no participant identifier.}
}
\value{
A dataset with subscale scores for the Family Food Behavior Survey
}
\description{
This function scores the Family Food Behvior Survey and provides subscale scores for the following behaviors: Maternal Control, Maternal Presences, Child Choice, and Organization
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'ffbs#' where # is the question number (1-20)
3) All questions must have the numeric value for the choice: 0 - Never True, 1 - Rarely True, 2 - Sometimes, 3 - Often True, 4 - Always True

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the ffbs with IDs
ffbs_score_data <- score_ffbs(ffbs_data, parID = 'ID')

\dontrun{
}

}
\references{
Baughcum, A. E., Powers, S. W., Johnson, S. B., Chamberlin, L. A., Deeks, C. M., Jain, A., & Whitaker, R. C. (2001). Maternal Feeding Practices and Beliefs and Their Relationships to Overweight in Early Childhood: Journal of Developmental & Behavioral Pediatrics, 22(6), 391–408. https://doi.org/10.1097/00004703-200112000-00007 (\href{https://pubmed.ncbi.nlm.nih.gov/11773804/}{PubMed})

McCurdy, K., & Gorman, K. S. (2010). Measuring family food environments in diverse families with young children. Appetite, 54(3), 615–618. https://doi.org/10.1016/j.appet.2010.03.004 (\href{https://pubmed.ncbi.nlm.nih.gov/20227449/}{PubMed})
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_parent_v2dat}}
}
