% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hfias.R
\name{score_hfias}
\alias{score_hfias}
\title{score_hfias: Scored data from the Household Food Insecurity Access Scale}
\usage{
score_hfias(hfias_data, parID)
}
\arguments{
\item{hfias_data}{a data.frame all items for the Household Food Insecurity Access Scale following the naming conventions described above}

\item{parID}{(optional) name of participant ID column in intake_data. If included the output dataset will be matched by parID, if not included the output dataset will be in the order of intake_data but will have no participant identifier.}
}
\value{
A dataset with scores for the Household Food Insecurity Access Scale
}
\description{
This function scores the Household Food Insecurity Access Scale and provides an overall HFIAS Score and Food Insecurity Status
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'hfias#' where # is question number (1-9 and 1a-9a).
3) Questions must have the numeric values for the choices. The options include:
3a) Yes - 1, No - 0, I don't know or Don't want to answer - 99
3b) Rarely - 1, Sometimes - 2, Often - 3, I don't know or Don't want to answer - 99

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the hfias with IDs
hfias_score_data <- score_hfias(hfias_data, parID = 'ID')


\dontrun{
}

}
\references{
Coates J, Swindale A, Bilinsky P. Household Food Insecurity Access Scale (HFIAS) for measurement of food access: indicator guide: version 3. Published online 2007.
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_parent_v4dat}}
}
