% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hfssm.R
\name{score_hfssm}
\alias{score_hfssm}
\title{score_hfssm: Scored data from the US Household Food Security Survey Module: Three Stage}
\usage{
score_hfssm(hfssm_data, parID)
}
\arguments{
\item{hfssm_data}{a data.frame all items for the US Household Food Security Survey Module: Three Stage following the naming conventions described above}
}
\value{
A dataset with subscale scores for the US Household Food Security Survey Module: Three Stage
}
\description{
This function scores the US Household Food Security Survey Module: Three Stage and provides subscale scores for the following: Household Food Security Scale, U.S. Adult Food Security Scale, Six-Item Food Security Scale, Children’s Food Security Scale.
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items for each module. Only complete modules will be completed.
2) The  columns/variables must match the following naming convention: 'hfssm_mod#' where mod is the module ('hh' - household, 'ad' - adult, 'ch' - child).
3) To compute modules, the following questions are required:
3a) Household Module - requires items hfssm_hh2 through hfssm_ch7
3b) Adult Module - requires items hfssm_hh2 through hfssm_ch7
3c) Six-Item Module - requires items hfssm_hh2 through hfssm_ad5a
3d) Child Module - requires items hfssm_ch2 through hfssm_ch7
4) Questions must have the numeric values for the choices. The options include:
4a) Yes - 1, No - 0, I don't know or Don't want to answer - 99
4b) Often True - 1, Sometimes True - 1, Never True - 0, I don't know or Don't want to answer - 99
4c) Almost Every Month - 1, Some Months but not Every Month - 1, Only 1 or 2 months - 0, I don't know or Don't want to answer - 99

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the hfssm with IDs
hfssm_score_data <- score_hfssm(hfssm_data, parID = 'ID')


\dontrun{
}

}
\references{
Bickel G, Nord M, Price C, Hamilton W, Cook J, others. Guide to measuring household food security, revised 2000. US Department of Agriculture, Food and Nutrition Service. Published online 2000:52.

Nord M. Measuring Children’s Food Security in US Households, 1995-99. US Department of Agriculture, Economic Research Service; 2002.
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_parent_v4dat}}
}
