% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_lbc.R
\name{score_lbc}
\alias{score_lbc}
\title{score_lbc: Scored data from the Lifestyle Behavior Checklist}
\usage{
score_lbc(lbc_data, study = "fbs", parID)
}
\arguments{
\item{lbc_data}{a data.frame all items for the Lifestyle Behavior Checklist following the naming conventions described above}

\item{study}{a string indicating which study collected the data. Currently, only option and default is 'fbs'. This parameter is included so this script can be adapted for future studies that collect all questions - FBS skipped first 5 questions}
}
\value{
A dataset with subscale scores for the Lifestyle Behavior Checklist
}
\description{
This function scores the Lifestyle Behavior Checklist and provides subscale scores for the following behaviors: Food-Related Misbehavior, Overeating, Emotions Related to Overweight, and Physical Activity. NOTE: for FBS, caution should be used in interpreting the Food-Related Misbehavior and Overeating subscales as this study is missing 3 (of 7) and 2 (of 7) questions for the respective subscales.
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'lbc#' where # is the question number (1-25; note - script will allow only 6-25 for study = 'fbs')
3) All questions must have the numeric value for the choice: 1 - Not At All, 2 - A Little (-), 3 - A Little (+), 4 - Somewhat, 5 - Much (-), 6 - Much (+), 7 - Very Much

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the lbc with IDs
lbc_score_data <- score_lbc(lbc_data, parID = 'ID')

study specified
lbc_score_data <- score_lbc(lbc_data, study = 'fbs', parID = 'ID')

\dontrun{
}

}
\references{
West F, Sanders MR. The Lifestyle Behaviour Checklist: A measure of weight-related problem behaviour in obese children. International Journal of Pediatric Obesity. 2009;4(4):266-273. doi:10.3109/17477160902811199

Subscales:
West F, Morawska A, Joughin K. The Lifestyle Behaviour Checklist: evaluation of the factor structure. Child: Care, Health and Development. 2010;36(4):508-515. doi:10.1111/j.1365-2214.2010.01074.x (\href{https://pubmed.ncbi.nlm.nih.gov/20337641/}{PubMed})
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_parent_v3dat}}
}
