% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pds.R
\name{score_pds}
\alias{score_pds}
\title{score_pds: Scored data from Parent- or Self-Report Pubertal Development Scale}
\usage{
score_pds(pds_data, respondent, male = 0, female = 1, parID)
}
\arguments{
\item{pds_data}{a data.frame with child sex and all questions from the Pubertal Development Scale. The required data to score the Pubertal Development Scale include sex, questions 1-3, and male/female specific questions 4-5. These questions must have the following names in the dataset: 'sex', 'pds_1', 'pds_2', 'pds_3', 'pds_4m', 'pds_5m', 'pds_4f', 'pds_5fa'. Note, as long as variable names match those listed, pds_data can include only data from male and/or female version scales.}

\item{respondent}{string to indicate if parent or child completed the Pubertal Development Scale. Must enter either 'parent' or 'child'}

\item{male}{(optional) value for male. Default value is male = 0}

\item{female}{(optional) value for female. Default value is female = 1}

\item{parID}{(optional) name of participant ID column in intake_data. If included the output dataset will be matched by parID, if not included the output dataset will be in the order of intake_data but will have no participant identifier.}
}
\value{
A dataset with a Pubertal Development Score and the Tanner Stage equivalents.
}
\description{
This function scores the the Child- and Parent-Pubertal Development Scale. The pubertal development score (pds_score) is an average of all questions, ignoring any 'I Don't Know' responses. Following the unpublished manuscript guidelines by Crockett, L. J. (1988), the Tanner Stage equivalent is computed by sex: Males - on the sum of the scores for body hair growth, voice change, and facial hair growth; Females - sum of scores for body hair growth and breast development along with the binary response to menarche (yes/no). For males, 1 of the 3 items is allowed to be 'I Don't Know' while for females, 1 of the 2 growth questions was allowed to be 'I Don't Know' but response to menarche was required.
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must included a 'sex' variable (defualt: male = 0, female = 1)
2) The following questions are required from the Pubertal Development Scale: questions 1-3 and the sex-specific versions of questions 4-5. Additional variables are allowed in the provided data so long as the above naming conventions below are followed for the questions necessary to score the assessment.
3) The data set columns/variables must match the following naming convention: 'sex', 'pds_1', 'pds_2', 'pds_3', 'pds_4m', 'pds_5m', 'pds_4f', 'pds_5fa'.
4) All variable should be converted to numeric values such that 1 = not started yet, 2 = barely started, 3 = definitely started, 4 = seems complete, and 99 = I Don't Know. For Female ('pds_5fa') question on menarche, the response can be coded as Yes = 1, No = 0.

Note, as long as variable names match those listed, pds_data can include only data from male and/or female version scales.
}
\examples{
#default male/female
pds_tanner_data <- score_pds(pds_data, respondent = 'parent', parID = 'ID')

#specify male/female
pds_tanner_data <- score_pds(pds_data, respondent = 'parent', male = 'male', female = 'female', parID = 'ID')

\dontrun{

# male and female specification must match the data in brief_data. Do not give the value label if pds_data has label attributes for sex.

#check attributes for sex
attributes(pds_data$sex)

#$labels
#Male Female
# 0      1

#with the above attributes, the following will not run as the data.frame contains 0's and 1's, not the labels
pds_tanner_data <- score_pds(pds_data, age_var = 'age', sex_var = 'sex', male = 'Male', female = 'Female')

}

}
\references{
Carskadon, Mary A., and Christine Acebo. A self-administered rating scale for pubertal development. Journal of Adolescent Health 14, no. 3 (1993): 190-195. https://doi.org/10.1016/1054-139X(93)90004-9 (\href{https://pubmed.ncbi.nlm.nih.gov/8323929/}{PubMed})

Crockett, L. J. Pubertal development scale: Pubertal categories. Unpublished manuscript (1988).
}
\seealso{
For the Food and Brain Study, raw data from Qualtrics was processed using \code{\link{util_fbs_parent_v1dat}}, \code{\link{util_fbs_child_v7dat}}, and \code{\link{util_fbs_child_v7dat_home}},
}
