% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_psi.R
\name{score_psi}
\alias{score_psi}
\title{score_psi: Score data from the Parenting Style Inventory-II}
\usage{
score_psi(psi_data, study, parID)
}
\arguments{
\item{psi_data}{a data.frame all items for the Parenting Style Inventory-II following the naming conventions described above}

\item{study}{a string indicating which study collected the data. Currently, only option and default is 'fbs'. This parameter is included so this script can be adapted for future studies that use published question 1 and 2 phrasing that need to be reversed scored.}

\item{parID}{(optional) name of participant ID column in intake_data. If included the output dataset will be matched by parID, if not included the output dataset will be in the order of intake_data but will have no participant identifier.}
}
\value{
A dataset with a score for the Parenting Style Inventory-II
}
\description{
This function scores ONLY the Responsiveness subscale of the Parenting Style Inventory-II. Questions were asked for both mom AND dad.
}
\details{
Note: for the Food and Brain Study, questions 1 and 2 were re-phrased so no reverse scoring was needed. Reverse scoring is included in the script for other studies that use the published question phrasing.

To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'psi_resp_mom#' and 'psi_resp_dad#' where # is the question number (1-5)
3) All questions must have the numeric value for the choice: 1 - Strongly Disagree, 2 - Disagree, 3 - I am in between, 4 - Agree, 5 - Strongly Agree

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the psi with IDs
psi_score_data <- score_psi(psi_data, study = 'fbs', parID = 'ID')

\dontrun{
# won't run without parameter study
psi_score_data <- score_psi(psi_data, parID = 'ID')
}

}
\references{
Darling N, Toyokawa T. Construction and validation of the parenting style inventory II (PSI-II). Unpublished manuscript. 1997;89. (https://www.researchgate.net/profile/Nancy-Darling/publication/341909949_Construction_and_Validation_of_the_Parenting_Style_Inventory_II_PSI-II/links/5ed9027792851c9c5e7bc63d/Construction-and-Validation-of-the-Parenting-Style-Inventory-II-PSI-II.pdf)
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_child_v4dat}} and \code{\link{util_fbs_child_v4dat_home}}
}
