% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pwlb.R
\name{score_pwlb}
\alias{score_pwlb}
\title{score_pwlb: Scored data from the Parent Weight-Loss Behavior Questionnaire}
\usage{
score_pwlb(pwlb_data, parID)
}
\arguments{
\item{pwlb_data}{a data.frame all items for the Parent Weight-Loss Behavior Questionnaire following the naming conventions described above}
}
\value{
A dataset with subscale scores for the Parent Weight-Loss Behavior Questionnaire
}
\description{
This function scores the Parent Weight-Loss Behavior Questionnaire and provides subscale scores for the following behaviors: Healthy Weight Control and Unhelathy Weight Control.
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'pwlb#' where # is the question number (1-29; only 1-24 are needed to compute scale as 25-29 are contextual questions)
3) Questions 1-24 must have the numeric value for the choices: 1 - Never, 2 - Rarely, 3 - Sometimes, 4 - Often, 5 - Always. Questions 25-29 are scored differently and are follow-up questions.

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the pwlb with IDs
pwlb_score_data <- score_pwlb(pwlb_data, parID = 'ID')


\dontrun{
}

}
\references{
Primary References for the Parent Weight-Loss Behavior Questionnaire and Scoring:
Savage JS, Birch LL. Patterns of weight control strategies predict differences in women’s 4 y weight gain. Obesity (Silver Spring). 2010;18(3):513-520. doi:10.1038/oby.2009.265 (\href{https://pubmed.ncbi.nlm.nih.gov/19696759/}{PubMed})

Measure Adapted from:
French SA, Perry CL, Leon GR, Fulkerson JA. Dieting behaviors and weight change history in female adolescents. Health Psychology. 1995;14(6):548-555. doi:http://dx.doi.org/10.1037/0278-6133.14.6.548 (\href{https://pubmed.ncbi.nlm.nih.gov/8565929/}{PubMed})
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_parent_v3dat}}
}
