% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_rcmas.R
\name{score_rcmas}
\alias{score_rcmas}
\title{score_rcmas: Score data from the Revised Children's Manifest Anxiety Scale}
\usage{
score_rcmas(rcmas_data, parID)
}
\arguments{
\item{rcmas_data}{a data.frame all items for the Revised Children's Manifest Anxiety Scale following the naming conventions described above}
}
\value{
A dataset with a score for the Revised Children's Manifest Anxiety Scale
}
\description{
This function scores the Revised Children's Manifest Anxiety Scale for total score and the following subscales: Physiological Manifestations, Worry and Oversensitivity, and Fear/Concentration
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items and child grade. If child grade is not included some variables will be exported with NA's.
2) The columns/variables must match the following naming convention: 'rcmas#' where # is the question number (1-37) and child's grade must be labeled 'grade'
3) All questions must have the numeric value for the choice:
0 - No, 1 - Yes

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the rcmas with IDs
rcmas_score_data <- score_rcmas(rcmas_data, parID = 'ID')

\dontrun{
}

}
\references{
Reynolds CR, Richmond BO. What I think and feel: A revised measure of children’s manifest anxiety. Journal of abnormal child psychology. 1978;6(2):271-280. (\href{https://pubmed.ncbi.nlm.nih.gov/670592/}{PubMed})

Reynolds CR, Richmond BO. Factor structure and construct validity of’What I think and feel’: The revised children’s manifest anxiety scale. Journal of personality assessment. 1979;43(3):281-283. (\href{https://pubmed.ncbi.nlm.nih.gov/469706/}{PubMed})

Stallard P, Velleman R, Langsford J, Baldwin S. Coping and psychological distress in children involved in road traffic accidents. British Journal of Clinical Psychology. 2001;40(2):197-208. (\href{https://pubmed.ncbi.nlm.nih.gov/11446241/}{PubMed})
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{qualtrics_child_v2dat}} and \code{\link{qualtrics_child_v2dat_home}}
}
