% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_risk.R
\name{score_risk}
\alias{score_risk}
\title{score_risk: Determine Risk Status of Child}
\usage{
score_risk(risk_data, respondent, parID)
}
\arguments{
\item{risk_data}{a data.frame all parent BMI data following the naming conventions described above}

\item{respondent}{a string indicating the variable name for the parent who completed the questionnaires and had height/weight measured at Visit 1}
}
\value{
A dataset with amount of MVPA calculated for each day and week. If sleep = TRUE, parent-reported sleep will also be included
}
\description{
This function determines the risk status of children based on parent BMI. Low risk: both parents BMI < 25; High risk: Mom BMI >= 30 and Dad BMI >=25. An error of +/- 0.5 BMI is allowed but only 1 parent can have BMI = 25.5 for Low Risk.
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include who completed the questionnaire (i.e., the respondent), the measured parent BMI, and both parents' reported BMIs.The respondent will have measured BMI at Visit 1 and will self-report the height and weight of the other parent.
2) The BMI columns/variables must match the following naming conventions - 'parent_bmi', 'sr_mom_bmi', and 'sr_dad_bmi'

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# risk categories
risk_score <- score_risk(risk_data, respondent = 'parent_respondent', parID = 'ID')

\dontrun{
}

}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_parent_v1dat}}, \code{\link{util_merge_v1}}
}
