% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_spsrq.R
\name{score_spsrq}
\alias{score_spsrq}
\title{score_spsrq: Scored data from the Sensitivity to Punishment and Sensitivity to Reward Questionnaire}
\usage{
score_spsrq(spsrq_data, parID)
}
\arguments{
\item{spsrq_data}{a data.frame all items for the Sensitivity to Punishment and Sensitivity to Reward Questionnaire following the naming conventions described above}

\item{parID}{(optional) name of participant ID column in intake_data. If included the output dataset will be matched by parID, if not included the output dataset will be in the order of intake_data but will have no participant identifier.}
}
\value{
A dataset with subscale scores for the Sensitivity to Punishment and Sensitivity to Reward Questionnaire
}
\description{
This function scores the Sensitivity to Punishment and Sensitivity to Reward Questionnaire and provides subscale scores for the following behaviors (2011; 48 item subscales): Fear/Shyness, Anxiety, Conflict Avoidance, Sensory Reward, Drive, Responsiveness to Social Approval, Impulsivity/Fun Seeking. The original 4 subcales (2004; 34 item subscales): Sensitivity to Punishment, Impulsivity/Fun Seeking, Drive, and Reward Responsiveness.
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'spsrq#' where # is the question number (1-48)
3) Questions 1-36 must have the numeric value for the choices: 1 - Strongly Disagree, 2 - Disagree, 3 - Neither Agree nor Disagree, 4 - Agree, 5 - Strongly Agree.

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the spsrq with IDs
spsrq_score_data <- score_spsrq(spsrq_data, parID = 'ID')


\dontrun{
}

}
\references{
Primary References for the Sensitivity to Punishment and Sensitivity to Reward Questionnaire and Scoring:
Updated/Revised Caregiver (2011 scoring; 48 items) - Colder CR, Trucco EM, Lopez HI, et al. Revised reinforcement sensitivity theory and laboratory assessment of BIS and BAS in children. Journal of Research in Personality. 2011;45(2):198-207. doi:10.1016/j.jrp.2011.01.005 (\href{https://pubmed.ncbi.nlm.nih.gov/21603055/}{PubMed})

Original Caregiver (2004 scoring; 34 items) - Colder CR, O’Connor RM. Gray’s Reinforcement Sensitivity Model and Child Psychopathology: Laboratory and Questionnaire Assessment of the BAS and BIS. J Abnorm Child Psychol. 2004;32(4):435-451. doi:10.1023/B:JACP.0000030296.54122.b6 (\href{https://pubmed.ncbi.nlm.nih.gov/15305548/}{PubMed})

Original/adult citation:
Torrubia R, Ávila C, Moltó J, Caseras X. The Sensitivity to Punishment and Sensitivity to Reward Questionnaire (SPSRQ) as a measure of Gray’s anxiety and impulsivity dimensions. Personality and Individual Differences. 2001;31(6):837-862. doi:10.1016/S0191-8869(00)00183-5
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_parent_v3dat}}
}
