% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_tesqe.R
\name{score_tesqe}
\alias{score_tesqe}
\title{score_tesqe: Score data from the Tempest Self-Regulation Questionnaire for Eating}
\usage{
score_tesqe(tesqe_data, parID)
}
\arguments{
\item{tesqe_data}{a data.frame all items for the Tempest Self-Regulation Questionnaire for Eating following the naming conventions described above}
}
\value{
A dataset with a score for the Tempest Self-Regulation Questionnaire for Eating
}
\description{
This function scores the Tempest Self-Regulation Questionnaire for Eating for the following subscales: Avoidance of Temptations, Controlling Temptations, Distraction, Suppression, Setting Goals and Rules, and Goal Deliberation
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'tesqe#' where # is the question number (1-24)
3) All questions must have the numeric value for the choice:
1 - Never, 2 - Rarely, 3 - Sometimes, 4 - Regularly, 5 - Always

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the tesqe with IDs
tesqe_score_data <- score_tesqe(tesqe_data, parID = 'ID')

\dontrun{
}

}
\references{
De Vet E, De Ridder D, Stok M, Brunso K, Baban A, Gaspar T. Assessing self-regulation strategies: development and validation of the tempest self-regulation questionnaire for eating (TESQ-E) in adolescents. Int J Behav Nutr Phys Act. 2014;11(1):106. doi:10.1186/s12966-014-0106-z
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_child_v2dat}} and \code{\link{util_fbs_child_v2dat_home}}
}
