% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_tfeq.R
\name{score_tfeq}
\alias{score_tfeq}
\title{score_tfeq: Scored data from the Three Factor Eating Questionnaire}
\usage{
score_tfeq(tfeq_data, parID)
}
\arguments{
\item{tfeq_data}{a data.frame all items for the Three Factor Eating Questionnaire following the naming conventions described above}
}
\value{
A dataset with subscale scores for the Three Factor Eating Questionnaire
}
\description{
This function scores the Three Factor Eating Questionnaire and provides subscale scores for the following behaviors: Cognitive Control of Eating Behaviors, Disinhibition of Control, and Susceptibility to Hunger.
}
\details{
Note - the factor structure of this questionnaire has been repeatedly questioned so may want to look further to decide on best approach for each study. The provided subscales are based on the canonical approach in the literature and should not be taken as a recommendation or 'best' factoring approach.

To use this function, the data must be prepared according to the following criteria:
1) The data must include all individual questionnaire items
2) The  columns/variables must match the following naming convention: 'tfeq#' where # is the question number (1-51)
3) Questions 1-36 must have the numeric value for the choices: 0 - False, 1 - True. Questions 37-51 must have numeric values 1-4 for the various responses with 1 corresponding to the least frequent/likely options and 4 corresponding to the most frequent/most likely option available. The option differ by question.

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the tfeq with IDs
tfeq_score_data <- score_tfeq(tfeq_data, parID = 'ID')


\dontrun{
}

}
\references{
Primary References for the Three Factor Eating Questionnaire and Scoring:
Stunkard AJ, Messick S. The three-factor eating questionnaire to measure dietary restraint, disinhibition and hunger. Journal of Psychosomatic Research. 1985;29(1):71-83. doi:10.1016/0022-3999(85)90010-8
}
\seealso{
Raw data from Qualtrics was processed using the following script: \code{\link{util_fbs_parent_v3dat}}
}
