% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fbs_child_v3dat_lab.R
\name{util_fbs_child_v3dat_lab}
\alias{util_fbs_child_v3dat_lab}
\title{util_fbs_child_v3dat_lab: Process raw qualtrics visit 3 data for the child}
\usage{
util_fbs_child_v3dat_lab(date_str, data_path)
}
\value{
A list containing: 1) data: data.frame with raw, cleaned data from child visit 3 Qualtrics and 2) dict: all variable descriptions
}
\description{
This function loads the .sav raw data file for the child visit 3 data that was collected via Qualtrics in the lab when the procedure was split due to covid and cleans the data. Cleaning the data involves:
1) extracting all variable descriptions,
2) selecting relevant data columns,
3) removing all practice events (e.g., 999)
4) re-ordering and re-name data columns
5) reformatting dates to be appropriate and computer readable: YYYY-MM-DD
6) re-calculate manual variables
7) re-ordering factor levels to start with value 0
8) random fixes to factor level names and variable descriptions
}
\details{
The databases MUST follow the naming convention: Child_V3_Lab_YYYY-MM-DD.sav
}
\examples{
#if in same working directory as data:
ch_v3_dat_lab <- util_fbs_child_v3dat_lab('2021-10-11')

\dontrun{
#date must be a string. The following will not run:
ch_v3_dat_lab <- util_fbs_child_v3dat_lab(2021-10-11)

#date must match the file name - for file named 'Child_V3_Lab_2021_09_16', the
following will not run:
ch_v3_dat_lab <- util_fbs_child_v3dat_lab('2021_10_11')
}


}
