% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fbs_child_v7dat_home.R
\name{util_fbs_child_v7dat_home}
\alias{util_fbs_child_v7dat_home}
\title{util_fbs_child_v7dat_home: Process raw qualtrics visit 7 data collected at home for the child}
\usage{
util_fbs_child_v7dat_home(date_str, data_path)
}
\arguments{
\item{date_str}{the date used in the name of the .sav file (e.g., for file 'Parent_V1_2021-10-11.sav', the string '2021-10-11' would be entered)}

\item{data_path}{(optional) the full path to the directory where the raw Qualtrics database are saved, EXCLUDING the .sav file name (e.g., '.../b-childfoodlab_Shared/Active_Studies/RO1_Brain_Mechanisms_IRB_5357/Participant_Data/untouchedRaw/Qualtrics_Raw/'). If NOT entered, will assume the database is saved in the working directory.}
}
\value{
A list containing: 1) data: data.frame with raw, cleaned data from child visit 7 Qualtrics; 2) dict: all variable descriptions; 3) pna_data: data.frame marking participants who 'prefered not to answer' (pna) specific questions; and 4) pna_dict: all variable descriptions for pna_data
}
\description{
This function loads the .sav raw data file for the child visit 7 data that was collected via Qualtrics at home when the protocol was split due to covid and cleans the data. Cleaning the data involves:
1) extracting all variable descriptions,
2) selecting relevant data columns,
3) removing all practice events (e.g., 999)
4) re-ordering and re-name data columns
5) reformatting dates to be appropriate and computer readable: YYYY-MM-DD
6) re-calculate manual variables
7) re-ordering factor levels
8) random fixes to factor level names and variable descriptions
}
\details{
The databases MUST follow the naming convention: Child_V7_YYYY-MM-DD.sav
}
\examples{
#if in same working directory as data:
ch_v7_dat_home <- util_fbs_child_v7dat_home('2021-10-11')

\dontrun{
#date must be a string. The following will not run:
ch_v7_dat_home <- util_fbs_child_v7dat_home(2021-10-11)

#date must match the file name - for file named 'Child_V7_Home_2021_09_16', the
following will not run:
ch_v7_dat_home <- util_fbs_child_v7dat_home('2021_10_11')
}


}
