% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fbs_merge_micro.R
\name{util_fbs_merge_micro}
\alias{util_fbs_merge_micro}
\title{util_fbs_merge_micro: Merge all microstructure data}
\usage{
util_fbs_merge_micro(data_path)
}
\arguments{
\item{data_path}{(optional) the full path to the directory where the raw Qualtrics database are saved, EXCLUDING the .sav file name (e.g., '.../b-childfoodlab_Shared/Active_Studies/RO1_Brain_Mechanisms_IRB_5357/Participant_Data/untouchedRaw/Qualtrics_Raw/'). If NOT entered, will assume the database is saved in the working directory.}
}
\value{
A list containing 2 data lists:
1) beh_wide contains - data: data.frame summary metrics in wide format by code, dict: data dictionary with variable descriptions
2) event contains - data: event data by coder and time, dict: data dictionary with variable descriptions
}
\description{
This function merges the raw microstructure data into databases
}
\details{
The databases MUST follow the naming convention: PS1_EventLogs_YYYY-MM-DD.txt, PS2_EventLogs_YYYY-MM-DD.txt, PS3_EventLogs_YYYY-MM-DD.txt, and PS4_EventLogs_YYYY-MM-DD.txt. The databases must all be in the SAME directory to be processed if the data_path is not entered and the directory organization does not follow the structure laid out in the DataManual.
}
\examples{
#if in same working directory as data with all data. Note - there is no need to add the COVID protocol (i.e., 'Home' or 'Lab') to file_pattern - these files will be search for automatically:
mciro_merged <- util_fbs_merge_micro(data_path = 'untouchedRaw/Microstructure_Raw/')

\dontrun{
}

}
\seealso{
Raw data is processed using \code{\link{util_fbs_microstructure}}
}
