% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fbs_merge_v6.R
\name{util_fbs_merge_v6}
\alias{util_fbs_merge_v6}
\title{util_fbs_merge_v6: Merge all Visit 6 raw Qualtrics databases and organize variable in database order}
\usage{
util_fbs_merge_v6(child_file_pattern, parent_file_pattern, data_path)
}
\arguments{
\item{child_file_pattern}{string with the pattern to search for to find the raw child data files. The pattern must contain respondent and visit number (e.g., for files from child visit 1, would enter 'Child_V1').}

\item{parent_file_pattern}{string with the pattern to search for to find the raw parent data files. The pattern must contain respondent and visit number (e.g., for files from parent visit 1, would enter 'Parent_V1').}

\item{data_path}{(optional) the full path to the directory where the raw Qualtrics database are saved, EXCLUDING the .sav file name (e.g., '.../b-childfoodlab_Shared/Active_Studies/RO1_Brain_Mechanisms_IRB_5357/Participant_Data/untouchedRaw/Qualtrics_Raw/'). If NOT entered, will assume the database is saved in the working directory.}
}
\value{
A list containing: 1) data: data.frame with raw, cleaned data from parent visit 6 Qualtrics and 2) dict: all variable descriptions
}
\description{
This function merges the following visit 6 raw data into a single database and organizes variables in database order: child visit 6, child visit 6-home, child visit 6-lab, and parent visit 6
}
\details{
The databases MUST follow the naming convention: Child_V6_YYYY-MM-DD.sav, Child_V6_Home_YYY-MM-DD.sav, Child_V6_Lab_YYY-MM-DD.sav, and Parent_V6_YYY-MM-DD.sav. The databases must all be in the SAME directory to be processed if the data_path is not entered and the directory organization does not follow the structure laid out in the DataManual.
}
\examples{
#if in same working directory as data with all data. Note - there is no need to add the COVID protocol (i.e., 'Home' or 'Lab') to file_pattern - these files will be search for automatically:
v6dat_scored <- util_fbs_merge_v6(child_file_pattern = 'Child_V6', parent_file_pattern = 'Parent_V6')

\dontrun{
#child_file_pattern and parent_file_pattern must be a strings. The following will not run:
v6dat_scored <- util_fbs_merge_v6(child_file_pattern = Child_V6, parent_file_pattern = Parent_V6)

# *_file_pattern must have the respondent ('Child') and visit number ('V6'). If just enter 'Child' or 'Parent', the script will not run because it will return multiple files for different child visits. The following will not run:
v6dat_scored <- util_fbs_merge_v6(child_file_pattern = 'Child', parent_file_pattern = 'Parent')
}

}
\seealso{
Raw data from Qualtrics is processed using the following scripts: \code{\link{util_fbs_child_v6dat}}, \code{\link{util_fbs_parent_v6dat}}. Visit 6 data is scored using the following scripts: \code{\link{score_mrivas}}.
}
