% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fbs_parent_v5dat.R
\name{util_fbs_parent_v5dat}
\alias{util_fbs_parent_v5dat}
\title{util_fbs_parent_v5dat: Process raw Qualtrics visit 5 data for the parent}
\usage{
util_fbs_parent_v5dat(file_pattern, data_path)
}
\arguments{
\item{file_pattern}{string with the pattern to search for to find the raw data file. The pattern must contain respondent and visit number (e.g., for files from parent visit 1, would enter 'Parent_V1').}

\item{data_path}{(optional) the full path to the directory where the raw Qualtrics database are saved, EXCLUDING the .sav file name (e.g., '.../b-childfoodlab_Shared/Active_Studies/RO1_Brain_Mechanisms_IRB_5357/Participant_Data/untouchedRaw/Qualtrics_Raw/'). If NOT entered, will assume the database is saved in the working directory.}
}
\value{
A list containing: 1) data: data.frame with raw, cleaned data from parent visit 5 Qualtrics;
 2) dict: all variable descriptions; 3) pna_data: data.frame marking participants who 'prefered not to answer' (pna) specific questions; and 4) pna_dict: all variable descriptions for pna_data
}
\description{
This function loads the .sav raw data file for the parent visit 5 data that was
collected via Qualtrics and cleans the data. Cleaning the data involves:
1) extracting all variable descriptions,
2) selecting relevant data columns,
3) removing all practice events (e.g., 999)
4) re-ordering and re-name data columns
5) general fixes to variable labels (remove ' - 1')
6) fix variables with 99 issue for 'prefer not to answer'
8) fix factor levels to match questionnaire scoring
}
\details{
The databases MUST follow the naming convention: Parent_V5_YYYY-MM-DD.sav
}
\examples{
#if in same working directory as data:
p_v5_dat <- util_fbs_parent_v5dat('Parent_V5')

\dontrun{
#file_pattern must be a string. The following will not run:
p_v5_dat <- util_fbs_parent_v5dat(Parent_V5)

#file_pattern must have the respondent ('Parent') and visit number ('V1'). If just enter 'Parent', the script will not run because it will return multiple files for different parent visits. The following will not run:
p_v5_dat <- util_fbs_parent_v5dat('Parent')
}


}
