% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_parent_v7dat.R
\name{util_parent_v7dat}
\alias{util_parent_v7dat}
\title{util_parent_v7dat: Process raw Qualtrics visit 7 data for the parent (1-year follow-up)}
\usage{
util_parent_v7dat(date_str, data_path)
}
\arguments{
\item{date_str}{the date used in the name of the .sav file (e.g., for file 'Parent_V1_2021-10-11.sav', the string '2021-10-11' would be entered)}

\item{data_path}{(optional) the full path to the directory where the raw Qualtrics database are saved, EXCLUDING the .sav file name (e.g., '.../b-childfoodlab_Shared/Active_Studies/RO1_Brain_Mechanisms_IRB_5357/Participant_Data/untouchedRaw/Qualtrics_Raw/'). If NOT entered, will assume the database is saved in the working directory.}
}
\value{
A list containing: 1) data: data.frame with raw, cleaned data from parent visit 7 Qualtrics; 2) dict: all variable descriptions; 3) pna_data: data.frame marking participants who 'prefered not to answer' (pna) specific questions; and 4) pna_dict: all variable descriptions for pna_data
}
\description{
This function loads the .sav raw data file for the parent visit 7 data that was collected via Qualtrics at 1-year follow-up and cleans the data. Cleaning the data involves:
1) extracting all variable descriptions,
2) selecting relevant data columns,
3) removing all practice events (e.g., 999)
4) re-ordering and re-name data columns
5) general fixes to variable labels (remove ' - 1')
6) fix variables with 99 issue for 'prefer not to answer'
7) reformatting dates to be appropriate and computer readable: YYYY-MM-DD
8) re-calculate manual variables
9) re-ordering factor levels to start with value 0
10) random fixes to factor level names and variable descriptions
}
\details{
The databases MUST follow the naming convention: Parent_V7_Home_YYYY-MM-DD.sav
}
\examples{
#if in same working directory as data:
p_v7_dat <- util_parent_v7dat('2021-10-11')

\dontrun{
#date must be a string. The following will not run:
p_v7_dat <- util_parent_v7dat(2021-10-11)

#date must match the file name - for file named 'Parent_V7_021_09_16', the
following will not run:
p_v7_dat <- util_parent_v7dat('2021_10_11')
}


}
