% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbs_kcal_intake.R
\name{fbs_kcal_intake}
\alias{fbs_kcal_intake}
\title{fbs_kcal_intake: Calculate the amount served and eaten in kcal for the Food and Brain Study}
\usage{
fbs_kcal_intake(intake_data, meal, parID)
}
\arguments{
\item{intake_data}{a data.frame including grams consumed for each food item following the naming conventions described above}

\item{meal}{string indicating which meals are included in intake_data: 'std_meal' - standard meal before EAH (visits 1 and 7); 'EAH' - eating in the absence of hunger paradigm (visits 1 and 7); and 'ps_meal' - portion size meal (visits 2-5)}

\item{parID}{(optional) name of participant ID column in intake_data. If included the output dataset will be matched by parID, if not included the output dataset will be in the order of intake_data but will have no participant identifier.}
}
\value{
The intake_data data.frame with kcal consumed added for each food item and total consumed in grams and kcal.
}
\description{
This function calculates the amount served and eaten in kcal for the Food and Brain Study using food energy density and measured gram weights. This is done for all meals included in the study: standard meal + eating in the absence of hunger (EAH) (visits 1 and 7) and the portion size meals (visits 2-5). The kcal consumed will be calculated for each food item and total grams and kcal consumed will be calculated IF all food items for a meal were included in the dataset for a participant. If a participant has missing data, total consumed will not be calculated for them.
}
\details{
To use this function, the data must be prepared according to the following criteria:
1) The data must include the amount consumed for each food item individually
2) The amount consumed columns/variables must match the following naming conventions for each food item - 'consumed_food_g' where 'food' is replaced with each food item
3) The following foods must be included for each meal
3a) standard meal: 'applesauce', 'carrot', 'cheese_sndwch', 'cookies', 'ham_sndwch, 'milk', 'pbj_sndwch', 'potatochip', 'turkey_sndwch', 'ketchup', 'mayo', 'mustard',
3b) EAH: 'brownies', 'cornchips', 'hersheys', icecream', 'oreos', 'popcorn', 'pretzels', 'skittles', 'starbursts', 'water'
3c) portion size meals: 'chkn_nug', 'mac_cheese', 'grapes', broccoli', ketchup', water'

Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{
# intake for standard meal
fbs_kcal_data <- fbs_kcal_intake(intake_data, meal = 'std_meal', parID = 'ID')

# intake for eah
fbs_kcal_data <- fbs_kcal_intake(intake_data, meal = 'EAH', parID = 'ID')

\dontrun{
# no meal specified
fbs_kcal_data <- fbs_kcal_intake(intake_data, parID = 'ID')
}

}
\seealso{
For the Food and Brain Study, the energy density of each food is in \link{fbs_ED.rda}.
Raw data from Qualtrics was processed using the following scripts (or there 'lab' equivalent):
 \code{\link{qualtrics_child_v1dat}}, \code{\link{qualtrics_child_v2dat}}, \code{\link{qualtrics_child_v3dat}}, \code{\link{qualtrics_child_v4dat}}, \code{\link{qualtrics_child_v5dat}}, \code{\link{qualtrics_child_v7dat}}
}
