% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fbs_merge_v7.R
\name{util_fbs_merge_v7}
\alias{util_fbs_merge_v7}
\title{util_fbs_merge_v7: Merge all Visit 7 raw Qualtrics databases and organize variable in database order}
\usage{
util_fbs_merge_v7(
  date_str,
  child_date_str,
  child_home_date_str,
  child_lab_date_str,
  parent_date_str,
  data_path
)
}
\arguments{
\item{date_str}{If ALL databases have the same  date used in the name of the .sav file, enter the date as a string (e.g., for file 'Parent_V1_2021-10-11.sav', the string '2021-10-11' would be entered)}

\item{child_date_str}{(optional) If the date string differs by file, enter the child full protocol date string}

\item{child_home_date_str}{(optional) If the date string differs by file, enter the child HOME protocol date string}

\item{child_lab_date_str}{(optional) If the date string differs by file, enter the child LAB protocol date string}

\item{parent_date_str}{(optional) If the date string differs by file, enter the parent date string}

\item{data_path}{(optional) the full path to the directory where the raw Qualtrics database are saved, EXCLUDING the .sav file name (e.g., '.../b-childfoodlab_Shared/Active_Studies/RO1_Brain_Mechanisms_IRB_5357/Participant_Data/untouchedRaw/Qualtrics_Raw/'). If NOT entered, will assume the database is saved in the working directory.}

\item{parent_home_date_str}{(optional) If the date string differs by file, enter the parent HOME protocol date string}
}
\value{
A list containing: 1) data: data.frame with raw, cleaned data from parent visit 7 Qualtrics; 2) dict: all variable descriptions; 3) pna_data: data.frame marking participants who 'preferred not to answer' (pna) specific questions; and 4) pna_dict: all variable descriptions for pna_data
}
\description{
This function merges the following visit 7 raw data into a single database and organizes variables in database order: child visit 7, child visit 7-home, child visit 7-lab, and parent visit 7
}
\details{
The databases MUST follow the naming convention: Child_V7_YYYY-MM-DD.sav, Child_V7_Home_YYY-MM-DD.sav, Child_V7_Lab_YYY-MM-DD.sav, and Parent_V7_YYY-MM-DD.sav. The databases must all be in the SAME directory to be processed if the data_path is not entered. If it is entered, it must follow
}
\examples{
#if in same working directory as data:
v7dat_scored <- util_fbs_merge_v7(date_str = '2021-10-11')

#if in same working directory as data and covid collecte data has different dates:
v7dat_scored <- util_fbs_merge_v7(child_date_str = '2021-10-11', child_home_date_str = '2021-9-15', child_lab_date_str = = '2021-9-15', parent_date_str = '2021-10-11')

\dontrun{
#date must be a string. The following will not run:
v7dat_scored <- util_fbs_merge_v7(2021-10-11)

#date must match the file name - for file named 'Child_V7_Home_2021_09_15', the
following will not run:
v7dat_scored <- util_fbs_merge_v7('2021_10_11')
}

}
\seealso{
Raw data from Qualtrics is processed using the following scripts: \code{\link{util_fbs_child_v7dat}}, \code{\link{util_fbs_child_v7dat_home}}, \code{\link{util_fbs_child_v7dat_lab}}, \code{\link{util_fbs_parent_v7dat}}. Visit 7 data is scored using the following scripts: \code{\link{score_ctc}}, \code{\link{score_audit}}
}
